requirements.txt
Flask==2.1.1
sentence-transformers==2.2.0
scikit-learn==1.0.2
numpy==1.21.2
pandas==1.3.3
spacy==3.1.0
joblib==1.1.0
